[redhat] Make Fedora Rawhide the default dist
With ARK and Fedora about to merge, make Fedora the default dist. ARK
will be built largely by automation, but Fedora will get a lot of
end-user usage, so default to what they'll likely want.

When ARK is forked, the default can be set to a sensible default for
RHEL 9.

Signed-off-by: Jeremy Cline <jcline@redhat.com>