New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PMU_SYSFS:

 This option enables sysfs file creation for PMU SPRs like MMCR* and PMC*.

 Symbol: PMU_SYSFS [=n]
 Type  : bool
 Defined at arch/powerpc/platforms/Kconfig.cputype:428
   Prompt: Create PMU SPRs sysfs file
   Location:
     -> Processor support

Cc: kernel-patches@redhat.com