[redhat] New configs in drivers/edac
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_EDAC_DMC520:

 Support for error detection and correction on the
 SoCs with ARM DMC-520 DRAM controller.

 Symbol: EDAC_DMC520 [=n]
 Type  : tristate
 Defined at drivers/edac/Kconfig:526
   Prompt: ARM DMC-520 ECC
   Depends on: EDAC [=y] && ARM64 [=y]
   Location:
     -> Device Drivers
       -> EDAC (Error Detection And Correction) reporting (EDAC [=y])

Enabled as a module at the request of Aristeu Rozanski
---

Cc: Aristeu Rozanski <aris@redhat.com>