[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_TEST_HMM:

 This is a pseudo device driver solely for testing HMM.
 Say M here if you want to build the HMM test module.
 Doing so will allow you to run tools/testing/selftest/vm/hmm-tests.

 If unsure, say N.

 Symbol: TEST_HMM [=n]
 Type  : tristate
 Defined at lib/Kconfig.debug:2221
   Prompt: Test HMM (Heterogeneous Memory Management)
   Depends on: RUNTIME_TESTING_MENU [=y] && TRANSPARENT_HUGEPAGE [=y] && DEVICE_PRIVATE [=y]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
 Selects: HMM_MIRROR [=n] && MMU_NOTIFIER [=y]

---

Signed-off-by: CKI@GitLab <cki-project@redhat.com>