[redhat] fix kernel changelog release mismatch in dist-release
The kernel changelog which is created also have a wrong release number eg.:

  ...
  +RHEL_RELEASE = 69
  --- a/redhat/kernel.changelog-8.99
  +++ b/redhat/kernel.changelog-8.99
  @@ -1,3 +1,7 @@
  +* Wed Nov 11 2020 Fedora Kernel Team <kernel-team@fedoraproject.org> [5.10.0-0.rc3.20201111giteccc87672492.68.test]
  ...

That's because of stale data/changelog which is created before the release
bump is done. We need to revert changes and recreate the changelog after
the diff-index check and release bump, so we get the updated number and do
not have a duplicate changelog entry.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>