[redhat] New configs in fs/cifs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CIFS_SWN_UPCALL:

 The Service Witness Protocol (SWN) is used to get notifications
 from a highly available server of resource state changes. This
 feature enables an upcall mechanism for CIFS which contacts a
 userspace daemon to establish the DCE/RPC connection to retrieve
 the cluster available interfaces and resource change notifications.
 If unsure, say Y.

 Symbol: CIFS_SWN_UPCALL [=n]
 Type  : bool
 Defined at fs/cifs/Kconfig:193
   Prompt: SWN feature support
   Depends on: NETWORK_FILESYSTEMS [=y] && CIFS [=m]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])
         -> SMB3 and CIFS support (advanced network filesystem) (CIFS [=m])

---

Cc: kernel-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>