[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HID_FT260:

 Provides I2C host adapter functionality over USB-HID through FT260
 device. The customizable USB descriptor fields are exposed as sysfs
 attributes.

 To compile this driver as a module, choose M here: the module
 will be called hid-ft260.

 Symbol: HID_FT260 [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:354
   Prompt: FTDI FT260 USB HID to I2C host support
   Depends on: INPUT [=y] && HID [=y] && USB_HID [=y] && HIDRAW [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> HID support
         -> HID bus support (HID [=y])
           -> Special HID drivers

---

Cc: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>