[redhat] New configs in drivers/reset
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_RESET_MCHP_SPARX5:

 This driver supports switch core reset for the Microchip Sparx5 SoC.

 Symbol: RESET_MCHP_SPARX5 [=n]
 Type  : bool
 Defined at drivers/reset/Kconfig:117
   Prompt: Microchip Sparx5 reset driver
   Depends on: RESET_CONTROLLER [=y] && (HAS_IOMEM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Reset Controller Support (RESET_CONTROLLER [=y])
 Selects: MFD_SYSCON [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>