[redhat] New configs in fs/pstore
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PSTORE_BLK [=n]
 Type  : tristate
 Defined at fs/pstore/Kconfig:172
   Prompt: Log panic/oops to a block device
   Depends on: MISC_FILESYSTEMS [=y] && PSTORE [=y] && BLOCK [=y]
   Location:
     Main menu
       -> File systems
         -> Miscellaneous filesystems (MISC_FILESYSTEMS [=y])
           -> Persistent store support (PSTORE [=y])
 Selects: PSTORE_ZONE [=n]

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: Al Stone <ahs3@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>