[redhat] New configs in drivers/i2c
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: I2C_DESIGNWARE_AMDPSP [=n]
 Type  : bool
 Defined at drivers/i2c/busses/Kconfig:565
   Prompt: AMD PSP I2C semaphore support
   Depends on: I2C [=y] && HAS_IOMEM [=y] && X86_MSR [=y] && ACPI [=y] && I2C_DESIGNWARE_PLATFORM [=m]
   Location:
     Main menu
       -> Device Drivers
         -> I2C support
           -> I2C support (I2C [=y])
             -> I2C Hardware Bus support
               -> Synopsys DesignWare Platform (I2C_DESIGNWARE_PLATFORM [=m])

---

Cc: Gopal Tiwari <gtiwari@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>