New configs in lib/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRC64_ROCKSOFT [=m]
 Type  : tristate
 Defined at lib/Kconfig:148
   Prompt: CRC calculation for the Rocksoft model CRC64
   Location:
     Main menu
       -> Library routines
 Selects: CRC64 [=m] && CRYPTO [=y] && CRYPTO_CRC64_ROCKSOFT [=m]
 Selected by [m]:
   - BLK_DEV_INTEGRITY_T10 [=m] && BLOCK [=y] && BLK_DEV_INTEGRITY [=y]